Move imageUtils to own file from dataUtils
Refs #397.
